Move drive commands to Drive command folder
